+++ lib/modules/manager/gomod/artifacts.ts
@@ -1,5 +1,4 @@
 import is from '@sindresorhus/is';
-import { diffLines } from 'diff';
 import semver from 'semver';
 import { quote } from 'shlex';
 import upath from 'upath';
@@ -26,7 +25,7 @@ import type {
   UpdateArtifactsConfig,
   UpdateArtifactsResult,
 } from '../types';
-import { parseLine } from './line-parser';
+import { getExtraDepsNotice } from './artifacts-extra';
 
 const { major, valid } = semver;
 
@@ -112,107 +111,6 @@ function useModcacherw(goVersion: string | undefined): boolean {
     !Number.isNaN(minor) &&
     (major > 1 || (major === 1 && minor >= 14))
   );
-}
-
-interface ExtraDep {
-  depName: string;
-  currentValue: string;
-  newValue: string;
-}
-
-function getExtraDeps(
-  goModBefore: string,
-  goModAfter: string,
-  excludeDeps: string[],
-): ExtraDep[] {
-  const result: ExtraDep[] = [];
-
-  const diff = diffLines(goModBefore, goModAfter, {
-    newlineIsToken: true,
-  });
-
-  const addDeps: Record<string, string> = {};
-  const rmDeps: Record<string, string> = {};
-
-  for (const { added, removed, value } of diff) {
-    if (!added && !removed) {
-      continue;
-    }
-
-    const res = parseLine(value);
-    if (!res) {
-      continue;
-    }
-
-    const { depName, currentValue } = res;
-    if (!depName || !currentValue) {
-      continue;
-    }
-
-    if (added) {
-      addDeps[depName] = currentValue;
-    } else {
-      rmDeps[depName] = currentValue;
-    }
-  }
-
-  for (const [depName, currentValue] of Object.entries(rmDeps)) {
-    if (excludeDeps.includes(depName)) {
-      continue;
-    }
-
-    const newValue = addDeps[depName];
-    if (newValue) {
-      result.push({
-        depName,
-        currentValue,
-        newValue,
-      });
-    }
-  }
-
-  return result;
-}
-
-function extraDepsTable(extraDeps: ExtraDep[]): string {
-  const tableLines: string[] = [];
-
-  tableLines.push('| **Package** | **Change** |');
-  tableLines.push('| ----------- | ---------- |');
-
-  for (const { depName, currentValue, newValue } of extraDeps) {
-    tableLines.push(
-      `| \`${depName}\` | \`${currentValue}\` -> \`${newValue}\` |`,
-    );
-  }
-
-  return tableLines.join('\n');
-}
-
-function getExtraDepsNotice(
-  goModBefore: string | null,
-  goModAfter: string | null,
-  excludeDeps: string[],
-): string | null {
-  if (!goModBefore || !goModAfter) {
-    return null;
-  }
-
-  const extraDeps = getExtraDeps(goModBefore, goModAfter, excludeDeps);
-  if (extraDeps.length === 0) {
-    return null;
-  }
-
-  const noticeLines: string[] = [];
-
-  noticeLines.push(
-    'In addition to the dependencies listed above, the following packages will also be updated:',
-  );
-  noticeLines.push('\n');
-  noticeLines.push(extraDepsTable(extraDeps));
-  noticeLines.push('\n');
-
-  return noticeLines.join('\n');
 }
 
 export async function updateArtifacts({
@@ -476,11 +374,11 @@ export async function updateArtifacts({
         },
       };
 
-      const excludeDeps = filterMap(updatedDeps, (dep) => dep?.depName);
+      const updatedDepNames = filterMap(updatedDeps, (dep) => dep?.depName);
       const extraDepsNotice = getExtraDepsNotice(
         newGoModContent,
         finalGoModContent,
-        excludeDeps,
+        updatedDepNames,
       );
 
       if (extraDepsNotice) {
