+++ lib/modules/manager/gomod/types.ts
@@ -4,3 +4,9 @@ export interface MultiLineParseResult {
   reachedLine: number;
   detectedDeps: PackageDependency[];
 }
+
+export interface ExtraDep {
+  depName: string;
+  currentValue: string;
+  newValue: string;
+}
